Extract auth module
